API-Builder: Added the episodes command. Added 4th level so we could have options for the episodes command. Ended up combining the season+episode to reduce on the complexity of the the api-builder logic plus since they are both required arguments it just worked out nice. This is purely a cosmetic api-builder thing, the builder still sends them segmented (and I think it should stay this way). episode: Changed 'fullPath' to 'full_path' to conform to our naming scheme. season_list: Converted the season list command to use sql query rather than calling the internal season list command so the builder was not dependent on the internal function.
Note:
Thought about breaking apart the rest of the optional options so that the user can see that you can mix-n-match options but get the same output.
Overall I don't think it's worth the extra logic/coding and I think it just makes the commands more complex/confusing then they need to be.
However, I think doing it for the History section the permutations of the options shows the flexibility while keeping it simple.