Fix image thumbnail aspect ratio
Resizing images using the thumbnail method preserves the aspect ratio,
which is then discarded by the HTML/CSS, which forces them to a specific
pixel size.  Images which aren't that pixel size (not all TVDB banners
are exactly the proscribed pixel size) get blurred by the web browser.

Also added a content-type header for the resized image and slightly
increased the JPEG quality.  Banners look way better as PNGs (JPEG blurs
text), but if this feature is added it should probably be a user pref
because PNGs are significantly larger.