Add check if allowed to change permissions before changing permissions
On linux when download user and sickbeard user are not the same,
moving files will result in download user still owning the files.
When SickBeard tries to chmod permissions it results in error,
because only root or owner are allowed to do this.
Check if user is allowed to change permissions.